More doc/cleanup
