+++ sdk/src/main/java/com/stytch/sdk/common/README.md
@@ -1,0 +1,3 @@
+# Package com.stytch.sdk.common
+This package contains resources shared between our B2B and Consumer SDKs, including helpful
+typealiases, utility functions, data models, and internal support code.
